Moving the samples and tools out of the cmd dir to have a more flat repo
