--- v2/cmd/samples/http/sender/main.go
@@ -1,42 +1,0 @@
-package main
-
-import (
-	"context"
-	"log"
-
-	cloudevents "github.com/cloudevents/sdk-go/v2"
-	cehttp "github.com/cloudevents/sdk-go/v2/protocol/http"
-)
-
-func main() {
-	ctx := cloudevents.ContextWithTarget(context.Background(), "http://localhost:8080/")
-
-	p, err := cloudevents.NewHTTP()
-	if err != nil {
-		log.Fatalf("failed to create protocol: %s", err.Error())
-	}
-
-	c, err := cloudevents.NewClient(p, cloudevents.WithTimeNow(), cloudevents.WithUUIDs())
-	if err != nil {
-		log.Fatalf("failed to create client, %v", err)
-	}
-
-	for i := 0; i < 10; i++ {
-		e := cloudevents.NewEvent()
-		e.SetType("com.cloudevents.sample.sent")
-		e.SetSource("https://github.com/cloudevents/sdk-go/v2/cmd/samples/httpb/sender")
-		_ = e.SetData(cloudevents.ApplicationJSON, map[string]interface{}{
-			"id":      i,
-			"message": "Hello, World!",
-		})
-
-		res := c.Send(ctx, e)
-		if !cloudevents.IsACK(res) {
-			log.Printf("Failed to send: %v", res)
-		} else {
-			var httpResult *cehttp.Result
-			cloudevents.ResultAs(res, &httpResult)
-			log.Printf("Sent %d with status code %d", i, httpResult.StatusCode)
-		}
-	}
-}
