--- v2/cmd/samples/kafka/sender/main.go
@@ -1,49 +1,0 @@
-package main
-
-import (
-	"context"
-	"log"
-
-	"github.com/Shopify/sarama"
-
-	cloudevents "github.com/cloudevents/sdk-go/v2"
-	"github.com/cloudevents/sdk-go/v2/protocol/kafka_sarama"
-)
-
-const (
-	count = 10
-)
-
-func main() {
-	saramaConfig := sarama.NewConfig()
-	saramaConfig.Version = sarama.V2_0_0_0
-
-	sender, err := kafka_sarama.NewSender([]string{"127.0.0.1:9092"}, saramaConfig, "test-topic")
-	if err != nil {
-		log.Fatalf("failed to create protocol: %s", err.Error())
-	}
-
-	defer sender.Close(context.Background())
-
-	c, err := cloudevents.NewClient(sender, cloudevents.WithTimeNow(), cloudevents.WithUUIDs())
-	if err != nil {
-		log.Fatalf("failed to create client, %v", err)
-	}
-
-	for i := 0; i < count; i++ {
-		e := cloudevents.NewEvent()
-		e.SetType("com.cloudevents.sample.sent")
-		e.SetSource("https://github.com/cloudevents/sdk-go/v2/cmd/samples/kafka/sender")
-		_ = e.SetData(cloudevents.ApplicationJSON, map[string]interface{}{
-			"id":      i,
-			"message": "Hello, World!",
-		})
-
-		err := c.Send(context.Background(), e)
-		if err != nil {
-			log.Printf("failed to send: %v", err)
-		} else {
-			log.Printf("sent: %d", i)
-		}
-	}
-}
