+++ v2/samples/amqp/receiver/main.go
