+++ v2/samples/http/receiver-direct/main.go
