+++ v2/samples/http/receiver-traced/main.go
