+++ v2/samples/kafka/receiver/main.go
