+++ v2/samples/message-interoperability/main.go
