+++ v2/samples/nats/receiver/main.go
