+++ v2/samples/stan/receiver/main.go
